removing endpoint section from api reference
